+++ packages/article-skeleton/src/article-skeleton.web.js
@@ -1,9 +1,11 @@
 import React, { Component, Fragment } from "react";
 import Ad, { AdComposer } from "@times-components/ad";
+import SaveAndShareBar from "@times-components/save-and-share-bar";
 import ArticleExtras from "@times-components/article-extras";
 import LazyLoad from "@times-components/lazy-load";
-import { spacing } from "@times-components/styleguide";
+import { spacing, breakpoints } from "@times-components/styleguide";
 import { withTrackScrollDepth } from "@times-components/tracking";
+import Context from "@times-components/context";
 import ArticleBody from "./article-body/article-body";
 import {
   articleSkeletonPropTypes,
@@ -14,7 +16,10 @@ import insertDropcapIntoAST from "./dropcap-util";
 import {
   BodyContainer,
   HeaderAdContainer,
-  MainContainer
+  MainContainer,
+  SaveShareContainer,
+  SaveShareRefContainer,
+  isStickyAllowed
 } from "./styles/responsive";
 import Head from "./head";
 
@@ -25,9 +30,11 @@ const adStyle = {
 class ArticleSkeleton extends Component {
   constructor(props) {
     super(props);
+    this.handleScroll = this.handleScroll.bind(this);
 
     this.state = {
-      articleWidth: null
+      articleWidth: null,
+      isSticky: false
     };
   }
 
@@ -36,6 +43,22 @@ class ArticleSkeleton extends Component {
     this.setState({
       articleWidth: this.node && this.node.clientWidth
     });
+    if (window) {
+      window.addEventListener("scroll", this.handleScroll);
+    }
+  }
+
+  componentWillUnmount() {
+    if (window) {
+      window.removeEventListener("scroll", this.handleScroll);
+    }
+  }
+
+  handleScroll() {
+    const offsetTop = this.sticky.getBoundingClientRect().top;
+    const isSticky = isStickyAllowed(breakpoints.huge) && offsetTop <= 1;
+
+    this.setState({ isSticky });
   }
 
   render() {
@@ -60,8 +83,9 @@ class ArticleSkeleton extends Component {
       template
     } = article;
 
-    const { articleWidth } = this.state;
+    const { articleWidth, isSticky } = this.state;
     const newContent = [...content];
+
     if (newContent && newContent.length > 0) {
       newContent[0] = insertDropcapIntoAST(
         newContent[0],
@@ -83,48 +107,76 @@ class ArticleSkeleton extends Component {
           this.node = node;
         }}
       >
-        <Head article={article} />
-        <AdComposer adConfig={adConfig}>
-          <LazyLoad rootMargin={spacing(10)} threshold={0.5}>
-            {({ observed, registerNode }) => (
-              <Fragment>
-                <HeaderAdContainer key="headerAd">
-                  <Ad
-                    contextUrl={url}
-                    section={section}
-                    slotName="header"
-                    style={adStyle}
-                  />
-                </HeaderAdContainer>
-                <MainContainer>
-                  <Header width={articleWidth} />
-                  <BodyContainer>
-                    <ArticleBody
-                      content={newContent}
-                      contextUrl={url}
-                      observed={observed}
-                      registerNode={registerNode}
-                      section={section}
-                    />
+        <Context.Consumer>
+          {({ user }) => (
+            <Fragment>
+              <Head article={article} />
+              <AdComposer adConfig={adConfig}>
+                <LazyLoad rootMargin={spacing(10)} threshold={0.5}>
+                  {({ observed, registerNode }) => (
+                    <Fragment>
+                      <HeaderAdContainer key="headerAd">
+                        <Ad
+                          contextUrl={url}
+                          section={section}
+                          slotName="header"
+                          style={adStyle}
+                        />
+                      </HeaderAdContainer>
+                      <MainContainer>
+                        <Header
+                          topicsAllowed={user.isLoggedIn}
+                          width={articleWidth}
+                        />
+                        <SaveShareContainer isSticky={isSticky}>
+                          <SaveShareRefContainer isSticky={isSticky}>
+                            <div
+                              ref={el => {
+                                this.sticky = el;
+                              }}
+                            >
+                              <SaveAndShareBar
+                                articleUrl={url}
+                                onCopyLink={() => {}}
+                                onSaveToMyArticles={() => {}}
+                                onShareOnEmail={() => {}}
+                              />
+                            </div>
+                          </SaveShareRefContainer>
+                        </SaveShareContainer>
+                        <BodyContainer>
+                          <ArticleBody
+                            content={newContent}
+                            contextUrl={url}
+                            observed={observed}
+                            registerNode={registerNode}
+                            section={section}
+                          />
 
-                    <ArticleExtras
-                      analyticsStream={analyticsStream}
-                      articleId={articleId}
-                      commentsEnabled={commentsEnabled}
-                      registerNode={registerNode}
-                      relatedArticleSlice={relatedArticleSlice}
-                      relatedArticlesVisible={
-                        !!observed.get("related-articles")
-                      }
-                      spotAccountId={spotAccountId}
-                      topics={topics}
-                    />
-                  </BodyContainer>
-                </MainContainer>
-              </Fragment>
-            )}
-          </LazyLoad>
-        </AdComposer>
+                          <ArticleExtras
+                            analyticsStream={analyticsStream}
+                            articleId={articleId}
+                            commentsAllowed={user.isLoggedIn}
+                            commentsEnabled={commentsEnabled}
+                            registerNode={registerNode}
+                            relatedArticleAllowed={user.isLoggedIn}
+                            relatedArticleSlice={relatedArticleSlice}
+                            relatedArticlesVisible={
+                              !!observed.get("related-articles")
+                            }
+                            spotAccountId={spotAccountId}
+                            topics={topics}
+                            topicsAllowed={user.isLoggedIn}
+                          />
+                        </BodyContainer>
+                      </MainContainer>
+                    </Fragment>
+                  )}
+                </LazyLoad>
+              </AdComposer>
+            </Fragment>
+          )}
+        </Context.Consumer>
       </article>
     );
   }