+++ packages/article/article.showcase.web.js
@@ -29,6 +29,21 @@ export default {
       name: "Article with template choice",
       platform: "web",
       type: "story"
+    },
+    {
+      component: ({ boolean, color, select }, { decorateAction }) =>
+        renderArticleConfig({
+          boolean,
+          color,
+          decorateAction,
+          hasScaling: false,
+          isTeaser: true,
+          link,
+          select
+        }),
+      name: "Article teaser",
+      platform: "web",
+      type: "story"
     }
   ],
   name: "Pages/Article"