+++ packages/edition-slices/src/tiles/tile-e/index.js
@@ -1,9 +1,8 @@
 import React from "react";
-import { View } from "react-native";
 import PropTypes from "prop-types";
+import Image from "@times-components/image";
 import {
   getTileImageUri,
-  TileImage,
   TileLink,
   TileSummary,
   withTileTracking
@@ -11,19 +10,17 @@ import {
 import styles from "./styles";
 
 const TileE = ({ onPress, tile }) => (
-  <TileLink onPress={onPress} tile={tile}>
-    <View style={styles.container}>
-      <TileImage
-        aspectRatio={4 / 5}
-        style={styles.imageContainer}
-        uri={getTileImageUri(tile, "crop45")}
-      />
-      <TileSummary
-        headlineStyle={styles.headline}
-        style={styles.summaryContainer}
-        tile={tile}
-      />
-    </View>
+  <TileLink onPress={onPress} style={styles.container} tile={tile}>
+    <Image
+      aspectRatio={4 / 5}
+      style={styles.imageContainer}
+      uri={getTileImageUri(tile, "crop45")}
+    />
+    <TileSummary
+      headlineStyle={styles.headline}
+      style={styles.summaryContainer}
+      tile={tile}
+    />
   </TileLink>
 );