+++ packages/edition-slices/src/tiles/tile-w/index.js
@@ -1,10 +1,9 @@
 import React from "react";
-import { View } from "react-native";
 import PropTypes from "prop-types";
+import Image from "@times-components/image";
 import {
   getTileImageUri,
   getTileSummary,
-  TileImage,
   TileLink,
   TileSummary,
   withTileTracking
@@ -12,21 +11,19 @@ import {
 import styles from "./styles";
 
 const TileW = ({ onPress, tile }) => (
-  <TileLink onPress={onPress} tile={tile}>
-    <View style={styles.container}>
-      <TileSummary
-        headlineStyle={styles.headline}
-        style={styles.summaryContainer}
-        summary={getTileSummary(tile, 125)}
-        summaryStyle={styles.summary}
-        tile={tile}
-      />
-      <TileImage
-        aspectRatio={16 / 9}
-        style={styles.imageContainer}
-        uri={getTileImageUri(tile, "crop169")}
-      />
-    </View>
+  <TileLink onPress={onPress} style={styles.container} tile={tile}>
+    <TileSummary
+      headlineStyle={styles.headline}
+      style={styles.summaryContainer}
+      summary={getTileSummary(tile, 125)}
+      summaryStyle={styles.summary}
+      tile={tile}
+    />
+    <Image
+      aspectRatio={16 / 9}
+      style={styles.imageContainer}
+      uri={getTileImageUri(tile, "crop169")}
+    />
   </TileLink>
 );