+++ packages/star-button/src/star-button.js
@@ -4,7 +4,7 @@ import Link from "@times-components/link";
 import { IconStar } from "@times-components/icons";
 import styles, { getTheme } from "./styles";
 
-const StarButton = ({ disabled, height, isDark, onPress, selected }) => {
+const StarButton = ({ disabled, height, isDark, onPress, selected, style }) => {
   const stars = getTheme({ isDark });
   const starState =
     (disabled && "disabled") || (selected && "selected") || "initial";
@@ -12,7 +12,11 @@ const StarButton = ({ disabled, height, isDark, onPress, selected }) => {
 
   return (
     // eslint-disable-next-line jsx-a11y/anchor-is-valid
-    <Link disabled={disabled} linkStyle={styles.container} onPress={onPress}>
+    <Link
+      disabled={disabled}
+      linkStyle={[styles.container, style]}
+      onPress={onPress}
+    >
       <IconStar
         fillColour={fillColour}
         height={height}
@@ -27,6 +31,7 @@ StarButton.propTypes = {
   disabled: PropTypes.bool,
   height: PropTypes.number,
   isDark: PropTypes.bool,
+  style: PropTypes.shape({}),
   onPress: PropTypes.func.isRequired,
   selected: PropTypes.bool
 };
@@ -35,6 +40,7 @@ StarButton.defaultProps = {
   disabled: false,
   height: 18,
   isDark: false,
+  style: null,
   selected: false
 };