+++ ios-app/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.130](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.129...@times-components/ios-app@0.7.130) (2019-05-08)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.129](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.128...@times-components/ios-app@0.7.129) (2019-05-07)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.128](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.127...@times-components/ios-app@0.7.128) (2019-05-07)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.127](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.126...@times-components/ios-app@0.7.127) (2019-05-03)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.126](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.125...@times-components/ios-app@0.7.126) (2019-05-03)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.125](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.124...@times-components/ios-app@0.7.125) (2019-05-03)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.124](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.123...@times-components/ios-app@0.7.124) (2019-05-03)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
 ## [0.7.123](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.122...@times-components/ios-app@0.7.123) (2019-05-01)
 
 **Note:** Version bump only for package @times-components/ios-app