+++ packages/article-flag/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.29](https://github.com/newsuk/times-components/compare/@times-components/article-flag@2.6.28...@times-components/article-flag@2.6.29) (2019-05-07)
+
+
+### Bug Fixes
+
+* remove ghost article flags view ([#1916](https://github.com/newsuk/times-components/issues/1916)) ([a10cea3](https://github.com/newsuk/times-components/commit/a10cea3))
+
+
+
+
+
 ## [2.6.28](https://github.com/newsuk/times-components/compare/@times-components/article-flag@2.6.27...@times-components/article-flag@2.6.28) (2019-05-01)
 
 **Note:** Version bump only for package @times-components/article-flag