+++ packages/pages/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.15](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.14...@times-components/pages@2.0.15) (2019-05-08)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.14](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.13...@times-components/pages@2.0.14) (2019-05-07)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.13](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.12...@times-components/pages@2.0.13) (2019-05-07)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.12](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.11...@times-components/pages@2.0.12) (2019-05-03)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.11](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.10...@times-components/pages@2.0.11) (2019-05-03)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.10](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.9...@times-components/pages@2.0.10) (2019-05-03)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.9](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.8...@times-components/pages@2.0.9) (2019-05-03)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
 ## [2.0.8](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.7...@times-components/pages@2.0.8) (2019-05-01)
 
 **Note:** Version bump only for package @times-components/pages