+++ packages/section/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.7](https://github.com/newsuk/times-components/compare/@times-components/section@0.13.6...@times-components/section@0.13.7) (2019-05-08)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
+## [0.13.6](https://github.com/newsuk/times-components/compare/@times-components/section@0.13.5...@times-components/section@0.13.6) (2019-05-07)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
+## [0.13.5](https://github.com/newsuk/times-components/compare/@times-components/section@0.13.4...@times-components/section@0.13.5) (2019-05-07)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
+## [0.13.4](https://github.com/newsuk/times-components/compare/@times-components/section@0.13.3...@times-components/section@0.13.4) (2019-05-03)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
+## [0.13.3](https://github.com/newsuk/times-components/compare/@times-components/section@0.13.2...@times-components/section@0.13.3) (2019-05-03)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
+## [0.13.2](https://github.com/newsuk/times-components/compare/@times-components/section@0.13.1...@times-components/section@0.13.2) (2019-05-03)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
+## [0.13.1](https://github.com/newsuk/times-components/compare/@times-components/section@0.13.0...@times-components/section@0.13.1) (2019-05-03)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
 # [0.13.0](https://github.com/newsuk/times-components/compare/@times-components/section@0.12.5...@times-components/section@0.13.0) (2019-05-01)